Code Refactor - Destructure into individual components - Test all the components
